+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This projec
 
 - Width of many graphs adjusted from 6 to 8 inches
 - Library overview: Added average track length, average play count, average file size, and total library size
+- Album graphs: Now include artist and italicized album name
 
 ## 1.0.0 - 2025-07-05